backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
rank_bm25
pypdf
